airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-sql==1.30.2
apache-airflow-providers-postgres==6.5.1
airflow-ai-sdk[gemini]==0.1.7
pydantic-ai-slim[google]==1.41.0
supabase==2.27.1
